+++ NEWS.md
@@ -1,5 +1,7 @@
 # broom (development version)
 
+* Added an `intercept` argument to `tidy.aov()`, a logical indicating whether to include information on the intercept as the first row of results (#1144 by `@souza-victor`).
+
 # broom 1.0.3
 
 * Addressed test failures on R-devel.